closes #1.
In traversing the inductive declaration's telescope, I wasn't being aggressive enough in applying whnf reduction.

This commit makes some other changes, primarily to relieve the stack pressure from reduction via whnf_core, and increasing the stack size of spawned threads, which are 2MB by default; I think the C++ default is 8MB, so that's what I'm using for now. The main thread is still at whatever the system default is (it's not chosen by rust).
There are some other changes that can be made to move to iteration instead of recursion in the future as well.

After looking over the inductive module again, (and after spending the last year or so working with these kinds of indirected data structures) I really want to take a second crack at it.
Refining the abstractions used (particularly in the inductive module) in preparation for the implementation of nested inductives (probably this summer) is high on the todo list.