Update test files for new export format.
Removes some export tests for size; Init, and other parts of it needed
for more complex constructions, have become larger over time. I'll have
to figure out a good workaround. Because the exporter requires a
nontrivial environment setup, it's not easy to generate them on the fly
e.g. via a script.